Upgraded 1.27
Used the really neat explanation by Bill the Lizard again. I considered
using the fermat stuff from previous examples but decided to just
implement it as a direct translation of the question.

Now I know what congruent to a modulo n means! The expmod version is a
lot faster which is nice.